InvoiceCoreProcessor/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "invoice_core_processor"
version = "0.1.0"
authors = [
    { name="Jules Coder", email="jules@example.com" },
]
description = "A multi-agent system for automated invoice processing."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "pydantic-settings",
    "langgraph",
    "psycopg2-binary",
    "motor",
    "grpcio",
    "grpcio-tools",
    "pdfplumber",
    "python-docx",
    "pytesseract",
    "easyocr",
    "python-dotenv",
    "loguru",
    "openai",
    "alembic",
]

[project.urls]
"Homepage" = "https://github.com/example/InvoiceCoreProcessor"
"Bug Tracker" = "https://github.com/example/InvoiceCoreProcessor/issues"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["invoice_core_processor"]
